pyproject.toml
[project]
name = "agentic-discord-moderation-bot"
version = "0.1.0"
description = "A Discord bot with agentic AI capabilities using LangChain and LangGraph"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "py-cord>=2.7.0",
    "langchain>=1.2.0",
    "langgraph>=1.1.0",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_discord_moderation_bot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"